Adding new component, CVBuilder, with data model
instead of using TaskManager component, that will
come deprecated in future, the CVBuilder will allow
the user to start building his own CV.

the route has been changed in Signin.js to route
to a new route called /cv-builder, that has been
added to App.js, this will use the CVBuilder components.

The Resume.js model is the basis of the full template
for each each user.

the add section still not working properly.