the Create new CV functionality has been added
now the user in his profile can see all his created
cv if there is any, and can enter a new name to create
a new cv. when he create a new cv, then the user will
be redirected to cv-builder/${his_given_cv_name}.

the name of cv has been added to Resume.js model.
the route in App.js cv-builder/:name, name represent
any name given as an argument to be shown on the url
of the web browser.

the CVBuilder.js now uses axios.put, that is better
to use than post when the user want to modify his data
so that the modification is done on backend side, instead
of doing it in frontend.