testing AWS to send autentication code on email
we tested the code for the user authentification,
but it is not working properly. Because we need to
make the registration twice to have the verification
email.
We can solve by sending the email twice with a for
low and we can make a button to resend the email.

Next Step(feature): If the user forget his password
we implement a possibility to create a new password
that it is related to this work (AWS).