new design with predefined templates is added
added the MyCVs components to be called from Header
to show the user previously created CVs, the Template1
represent a pre defined Template that will use as
default the data from the SampleData1.js when the
user create a new CV, but when the user has already
a created and modified cv, then his data will be
passed to Template1 as prop.

the css files were created for better testig, but
are not okay.