Improvements on contacts, adding icons to preveiw
the dropdown list of icons has been added also for the
already added contacts, in case for changes or updates
the icons added on Template1Preview component. allowing the user
to click on website links and email links to open them in a new tab
this is not tested yet.